Removing all use of error suppression (the @ symbol or "Error Control Operators" as it is called they are called in the PHP manual) in the phpmyadmin codebase.
I have already posted my reasoning in the php manual (also reprinted below) as to why
having this in php code is a very bad idea:

http://www.php.net/manual/en/language.operators.errorcontrol.php#90987
==========
Error suppression should be avoided if possible as it doesn't just suppress the error that
you are trying to stop, but will also suppress errors that you didn't predict would ever
occur. This will make debugging a nightmare.

It is far better to test for the condition that you know will cause an error before preceding
to run the code. This way only the error that you know about will be suppressed and not
all future errors associated with that piece of code.

There may be a good reason for using outright error suppression in favor of the method I
have suggested, however in the many years I've spent programming web apps I've yet to
come across a situation where it was a good solution. The examples given on this manual
page are certainly not situations where the error control operator should be used.
==========

This situation is quite bad in phpmyadmin as it is used all throughout the codebase and
without documentation as to why the supression was added. It appears though that many
of the places where it was used was not just ignorance of the better method, but was flat
out laziness because the person writing the code didn't wish to do propper error checking
and instead just decided to hide errors.

The complete missuse of error supression in areas where it was just down to lazy
programming and the lack of documentation of why it was used in other areas has lead me
to not trust it's inclusion anywhere was for a good reason. Because of this I have  just
completely removed all supression and not even attempted to address any issue it may have
been hiding. These issues will be much better addressed one at a time as they are discovered
from use of the code base since unhiding of the errors.